Graphics_Effect 编码告警修复 Signed-off-by: wangpengbo<wangpengbo23@h-partners.com>

+++ src/ge_color_gradient_shader_filter.cpp
@@ -103,7 +103,7 @@ bool GEColorGradientShaderFilter::CheckInParams(float* color, float* position, f
         return false;
     }
 
-    int arraySize = strengths_.size();
+    int arraySize = static_cast<int>(strengths_.size());
     if (!color || !position || !strength || tupleSize < arraySize) {
         LOGE("GEColorGradientShaderFilter::CheckInParams array size error\n");
         return false;
